@@ -1,16 +1,13 @@
 package tests
 
 import (
-	"fmt"
-	"io/ioutil"
+	"github.com/lets-dev-it-out/go-jsonschema/pkg/generator"
+	"github.com/stretchr/testify/require"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
-
-	"github.com/lets-dev-it-out/go-jsonschema/pkg/generator"
 )
 
 var basicConfig = generator.Config{
@@ -80,7 +77,7 @@ func TestBooleanAsSchema(t *testing.T) {
 }
 
 func testExamples(t *testing.T, cfg generator.Config, dataDir string) {
-	fileInfos, err := ioutil.ReadDir(dataDir)
+	fileInfos, err := os.ReadDir(dataDir)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -121,20 +118,19 @@ func testExampleFile(t *testing.T, cfg generator.Config, fileName string) {
 			goldenFileName := filepath.Join(filepath.Dir(fileName), outputName)
 			t.Logf("Using golden data in %s", mustAbs(goldenFileName))
 
-			goldenData, err := ioutil.ReadFile(goldenFileName)
+			goldenData, err := os.ReadFile(goldenFileName)
 			if err != nil {
 				if !os.IsNotExist(err) {
 					t.Fatal(err)
 				}
 				goldenData = source
 				t.Log("File does not exist; creating it")
-				if err = ioutil.WriteFile(goldenFileName, goldenData, 0655); err != nil {
+				if err = os.WriteFile(goldenFileName, goldenData, 0655); err != nil {
 					t.Fatal(err)
 				}
 			}
-			if diff, ok := diffStrings(t, string(goldenData), string(source)); !ok {
-				t.Fatal(fmt.Sprintf("Contents different (left is expected, right is actual):\n%s", *diff))
-			}
+
+			require.Equal(t, string(goldenData), string(source))
 		}
 	})
 }
@@ -149,37 +145,6 @@ func testFailingExampleFile(t *testing.T, cfg generator.Config, fileName string)
 			t.Fatal("Expected test to fail")
 		}
 	})
-}
-
-func diffStrings(t *testing.T, expected, actual string) (*string, bool) {
-	if actual == expected {
-		return nil, true
-	}
-
-	dir, err := ioutil.TempDir("", "test")
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	defer func() {
-		_ = os.RemoveAll(dir)
-	}()
-
-	if err := ioutil.WriteFile(fmt.Sprintf("%s/expected", dir), []byte(expected), 0644); err != nil {
-		t.Fatal(err.Error())
-	}
-	if err := ioutil.WriteFile(fmt.Sprintf("%s/actual", dir), []byte(actual), 0644); err != nil {
-		t.Fatal(err.Error())
-	}
-
-	out, err := exec.Command("diff", "--side-by-side",
-		fmt.Sprintf("%s/expected", dir),
-		fmt.Sprintf("%s/actual", dir)).Output()
-	if _, ok := err.(*exec.ExitError); !ok {
-		t.Fatal(err.Error())
-	}
-
-	diff := string(out)
-	return &diff, false
 }
 
 func titleFromFileName(fileName string) string {
