@@ -1,8 +1,8 @@
-// Package schema defines JSON schema types.
+// Package schemas defines JSON schema types.
 //
 // Code borrowed from https://github.com/alecthomas/jsonschema/
 //
-// Copyright (C) 2014 Alec Thomas
+// # Copyright (C) 2014 Alec Thomas
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
